Add pacwall timer
Pacwall is a gimmicky script that creates a graph of installed packages
to use as a desktop background. It would make sense to run it every time
we update, but it's kinda slow and it would make the process feel laggy.
Using a timer instead makes it seamless.

It is worth exploring whether a pacman hook should call the service
rather than the actual script to make it a non-blocking call.