mkinitcpio: Add zfs to HOOKS
This solves what I assume is a race condition where the zfs filesystems
would fail to mount because the module had not been inserted into the
kernel yet. Putting it in the initramfs means that it is already present
before systemd even starts.

Note the reordering of the keyboard module. This is recommended by the
arch wiki in order to be able to deal with possible errors caused by
zfs. I'm not sure this is actually useful if you don't use zfs as your /
but I don't see the harm in doing it anyway.