Add pacman hook to rehash zsh's cache
zsh caches the contents of /usr/bin/ for completion purposes. This poses
the issue of the completion suggestions being out-of-date after a
command that modifies the contents of /usr/bin/, which normally is any
pacman command that installs, removes, or upgrades a packages.

The mechanism of rehashing is this: after every executed command, zsh
will check /var/cache/zsh/pacman for its modification date. If the zsh
cache is older than the file, it will trigger a rehash, otherwise no
action is taken. On pacman's side, the hook writes the aforementioned
file after every install/remove/upgrade operation.

The downside of this approach is that open zsh sessions other than the
one performing the upgrade will not rehash until after at least one
command is executed (which could simply be an empty command). An
alternative exists where pacman can send a signal via `kill` to all
running zsh processes. zsh can trap that signal and perform the rehash
then. This makes it more "live" and also avoid unnecessary checks for
the rehashes. The downside is that we have to define the signal handler
and if it's missing it will terminate instead. On the contrary, the
current method will fail cleanly if zshrc is not sourced. I'm thinking
of possible fail cases to evaluate whether the signal approach is robust
enought to use instead of this.

See here for the sources of both methods:
https://wiki.archlinux.org/index.php/Zsh#Persistent_rehash