Disable initramfs compression
The images are pretty small so even with a good compression ratio the
savings are in the order of 10MB. I don't know if there is any
noticeable impact in image generation and extraction times, but it isn't
even worth looking into to save so little space. Just disable it.