pacman: Keep more packages when upgrading
It's not like I'm often tight on space, and this makes it less likely to
remove a version that is later required for downgrading. Keep in mind
that we can always use online archives in a pinch, so the point is not
to keep as many packages as possible, just enough to make it very likely
that we have the one we need.