sysctl: Remove vm.swappiness
The low swappiness value is based on outdated advice, and possibly
hasn't been relevant since HDDs. See [this][1] article for a good modern
explanation on both the importance of swap and of a high enough
swappiness value. According to the article, it's likely that a value of
100 would be optimal for my setup, but the default of 60 is probably
good enough and swap is so rarely needed that there's not much point to
administering a different configuration.

[1]: https://chrisdown.name/2018/01/02/in-defence-of-swap.html